+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Building the searchable options: `Unable to create shared archive file $IDE_CACHE_DIR/pycharm243.18137.19.jsa: (No such file or directory).`
 - Compatibility with Gradle dependency verification. Previously it was failing with `Failed to create MD5 hash for file`.
 - Rework how the IDEs from Plugin Verification are resolved. JetBrains/intellij-platform-gradle-plugin#1784
+- Exclude `kotlin-stdlib` and `kotlinx-coroutines` transitive dependencies in various variants from IntelliJ Platform dependencies. JetBrains/intellij-platform-gradle-plugin#1817
 
 ## [2.1.0]